thin: add supporting infrastructure for tracking snapshot chain even if there's a gap
Thin snapshots can be chained. If we remove LVs in the middle of the
chain, we would still like to know what was the original chain and
hence we'd like to know from which thin LVs an LV descended.

Normally we track only direct origin through seg->origin pointer
(snapshot -> origin path) and through lv->segs_using_this_lv for
the other way round (origin -> snapshots path).

This patch add seg->indirect_origin and lv->indirect_segs_using_this_lv
to track the dependency even if the chain is broken somewhere in the
middle. Actual implementation of the removal and proper setting
of the new fields and structures will be subject to subsequent
patches - this patch adds only the infrastructure for making
this extended dependency tracking possible.

For example, let's consider this thin snapshot chain:

  LV1 -- LV2 -- LV3 -- LV4

After removing LV 2, we end up with:

  LV1           LV3 -- LV4

This patch adds support for recording the gap that resulted from
removing the LV in thin snapshot chain (removed LVs are prefixed
with '!' here):

    ____________
   /            \
  LV1    !LV2    LV3 -- LV4

The !LV2 is the removed LV - recorded in lv->removed_ancestors
list (LV3 holds this list in this situation). The LV3 also records
LV1 as its indirect origin.

After removing LV3, we end up with:

   _____________________
  /                     \
 LV1    !LV2     !LV3    LV4

In terms of the new fields/structures, we end up with:

  LV4->removed_ancestor_lv_names = [ "LV2", "LV3" ]
  first_seg(LV4)->origin = NULL
  first_seg(LV4)->indirect_origin = LV1
  LV1->indirect_segs_using_this_lv = [ first_seg(LV4) ]

====

A little bit more complex example:

  LV1 -- LV2 -- LV3 -- LV4
            \
             --------------- LV5 -- LV6

Then after removing the LV2, we end up with:

     _____________________
    /__________           \
   //          \           \
  LV1    !LV2   LV3 -- LV4  \
                             \
                              LV5 -- LV6

Then after removing the LV5, we end up with:
     ____________________________
    /__________                  \
   //          \                  \
  LV1    !LV2   LV3 -- LV4         \
                                    \
                             !LV5    LV6

  LV3->removed_ancestor_lv_names = [ "LV2" ]
  first_seg(LV3)->origin = NULL
  first_seg(LV3)->indirect_origin = LV1

  LV6->removed_ancestor_lv_names = [ "LV5" ]
  first_seg(LV6)->origin = NULL
  first_seg(LV6)->indirect_origin = LV1

  LV1->indirect_segs_using_this_lv = [ first_seg(LV3), first_seg(LV6) ]