select: add support for range reserved values and flagging named-only values
This patch allows for registration and then recognition of reserved
values which are ranges, so they're composed of two values actually
to denote the lower and upper bound for the range (stored as an array
with exactly two items to define the boundaries).

Also, this patch allows for flagging reserved values as named-only
which means that such values are not strictly reserved. The strictly
reserved values are reserved values as originally designed before.

Where distinction between strictly-reserved and named-only values
is clearly visible is with comparisons. Normally, strictly reserved
value is not accounted for if we do "greater than" or "lower than"
comparisons, for example:

  1  2  3 ....
     |
    abc

 - we have "abc" as reserved value for field with value "2"
 - the value reported for the field is "abc" (or "2", it doesn't matter here)
 - the selection we're processing is -S 'field < abc'
 - the result of the selection gives nothing as "abc" is strictly
   reserved value (bound to "2") and there's no order defined for
   it and it would only match if we directly compared the value
   (so -S 'field = abc' would match)

With named-only values, the "abc" is named-only value for "2",
so selection -S 'field < abc" is the same as using -S 'field < 2',
hence giving result of "1". The "abc" is just a name for some value
so the value or its assigned name can be used equally.