report: add support for string list decorations (prefix and suffix for list items)
This patch extends string list reporting in libdevmapper so that list item
decorations (prefixes and suffixes) are automatically recognized. When
such decoration is recognized, it's not accounted for while processing
selection over the item.

For example, if "[" and "]" is recognized as item decoration, then
"[abc]" is equal to "abc" when it comes to selection. This allows
for string list items to be reported with arbitrary marks or flags
(by means of a prefix or suffix) while selection criteria still
applies for such items.