lvmlockd: move sanlock lease allocation
The lockd_init_lv() function no longer calls lvmlockd
to ask for the sanlock lease to be allocated for the LV,
i.e. lockd_init_lv_args().

Instead, it waits until vg_write() to do that, so the sanlock
lease does not need to be freed in all the cases where lvcreate
may quit before actually finishing.