toolcontext: move lvmetad, lvmpolld and filter init to lvm_run_command and _lvm_init
Move lvmetad, lvmpolld and filter initialization out of
create_toolcontext to lvm_run_command (for LVM commands) and
to _lvm_init (for lvm2app).

This makes it possible to avoid this initialization in
case we're processing a simple command that doesn't process
metadata in any way (e.g. lvmconfig).

These simple commands do not need daemons (lvmetad, lvmpolld,
lvmlockd, clvmd) running and they do not need to communicate with
them. So we don't need to initialize them uselessly.