lvmcache: add lock_type to VG summary and info structs
vgsummary information contains provisional VG information
that is obtained without holding the VG lock.  This info
can be used to lock the VG, and then read it with vg_read().
After the VG is read properly, the vgsummary info should
be verified.

Add the VG lock_type to the vgsummary so that it can be
used to decide whether locks are likely to exist for a
VG before actually locking and reading the VG.  After
reading the VG, locking decisions based on the lock_type
would be verified by looking at the actual lock_type from
the VG metadata.