commands: lvremove: recognize -r|--removed switch
When lvremove is used together with -r|--removed switch, dead LVs
are processed as well. This allows for removing dead LVs from history.

For example, we have lvol1 --> lvol2 --> lvol3 thin snapshot chain at
the beginning:

$ lvs -o name,name_removed,full_ancestors,full_descendants -O removed vg
  LV    RLV  FAncestors   FDescendants
  pool
  lvol1                   lvol2,lvol3
  lvol2      lvol1        lvol3
  lvol3      lvol2,lvol1

Then we remove the lvol2 which is then recorded as dead LV -rlvol0:

$ lvremove -ff vg/lvol2
  Logical volume "lvol2" successfully removed

$ lvs -r -o name,name_removed,full_ancestors,full_descendants -O removed vg
  LV     RLV     FAncestors    FDescendants
  pool
  lvol1                        -rlvol0,lvol3
  lvol3          -rlvol0,lvol1
  -lvol2 -rlvol0 lvol1         lvol3

Without the -r|--removed switch, removed entities are not processed and
they're simply not visible to the command:

$ lvremove -ff vg/-rlvol0
  Failed to find logical volume "vg/-rlvol0"

With lvremove used together with the -r|--removed switch, we can removed
items from history:

$ lvremove -r -ff vg/-rlvol0
  Dead logical volume "rlvol0" successfully removed

The indirect origin/indirect users relation is still kept though, we
just don't see the history of what was in between there before:

$ lvs -r -o name,name_removed,full_ancestors,full_descendants -O removed vg
  LV    RLV  FAncestors   FDescendants
  pool
  lvol1                   lvol3
  lvol3      lvol1