lvmetad: fix update_metadata with new vgid
When vgchange -u changes the vgid in an existing VG,
it invokes update_metadata() in lvmetad with the existing
VG name but a new vgid.  update_metadata() doesn't recognize
this condition, and ends up with in correct state, e.g.
the existing name is missing in the vgname_to_vgid hash table.

This fixes update_metadata() to recognize when an existing VG
name is being updated with a new vgid, and in that case first
calls remove_metadata() on the old vgid, before continuing
with update_metadata() which will add the updated VG as if it
were new.

One known shortcoming of this approach is that the VG is
missing from lvmetad hash tables completely in the short time
between the remove and adding it again.  Other commands that
look up the VG in that window will not find it.