Strictly ignore duplicate PVs
In commands and lvmetad, if a pvid already exists on
one device, and a second device appears with the same
pvid, then ignore the second device.

Previously, commands and lvmetad attempted to juggle both
devices for the same PV.

Commands using lvmetad will no longer see persistent
warnings about duplicates because they won't be seen.

In 'pvs -a' or 'pvs dev', an ignored duplicate device
now looks like a non-PV, and the chosen device in a
duplicate pair looks like the only device for the PV.