pvscan: --cache should scan all devices
When running pvscan --cache, the intention is to scan
all devices, i.e. the equivalent of running
'pvscan --cache dev' for each device.  But, some devices
holding duplicate PVs were being skipped, and lvmetad
was not being told about them.

This was a side effect of the fact that the lvmcache
layer is used internally during scanning (it's used
to transfer info between the different stages of the
command: label reading, vg reading, sending the vg.)

An unwanted side effect of using lvmcache is that it
eliminates duplicate PVs if it sees them.  This
prevented the command from sending some duplicate PVs
to lvmetad (when the first duplicate was preferred
over the second.)

The fix is to drop the lvmcache state between scanning
each PV.  There is no need to keep any cache state between
each PV being scanned (there's nothing being reused), and
the unneeded state is causing lvmcache to eliminate
duplicates which is harmful for pvscan --cache.

Also, remove clvm locking from pvscan_lvmetad.  This is
not needed since lvmetad cannot be used with clvm.  It
was interfering with clearing the cache.