lvmetad: refresh content at command start
At the start of the command, check if the lvmetad
content needs updating (token not matching), and
if so do the scan and update it.  This gets the
updating done right at the beginning, before
command processing is started. The early scan
already exists for the foreign VGs case, so this
is added to it.

Previously, the command would be part way through
its own processing before discovering via lvmetad_send
that lvmetad needed updating.  It's not good to
interrupt a command, insert a 'pvscan --cache', then
continue on with the original command.  The pvscan
is likely to interfere with the original command.