notify: change the approach to notifications
While the command runs, it just sets cmd flags
if a vg is written or a pv is written.

When the command is done working and after all
its locks are released, it checks if either flag
is set, and if so, connects to the dbus daemon
and sends one signal.

The signal indicates if a vg was written or a pv
was written.