test: test status reporting
For example:

$ lvs -J
  {
  "report": [
  {"lv_name":"root", "vg_name":"fedora", "lv_attr":"-wi-ao----","lv_size":"19.00g"},
  {"lv_name":"swap", "vg_name":"fedora", "lv_attr":"-wi-ao----","lv_size":"500.00m"},
  {"lv_name":"lvol0", "vg_name":"vg", "lv_attr":"-wi-a-----","lv_size":"32.00m"}
  ]
  ,
  "status": [
  {"status_type":"0", "status_message":"Test"}
  ]
  }

I'll do better indentation later, it's not important now.

The status used above is just an example, we need to agree on
exact status fields to be displayed - this is supposed to report
various status collected from process_each_* fns - hence it will be
possible to parse easily which items were processed successfully
and which not (for example, which VG/LVs were missing/nonexistent or
any other problems hit during command execution).